Added a low of new and interesting features.
From commit cbaf496:
- Added Wallet definition on its proper module;
- Added Wallet generation algorithm, much akin to Bitcoin (not secure
yet).

From commit b90d75e:
- Fixed bug on wallets lookup on default folder. Default folder is now
the folder .lulzcoin, created on the binary's run directory (and also
why it is blacklisted in .gitignore);
- Prepared to encrypt serialized wallet;
- Added a Wallet::load method to load a wallet from an arbitrary path
on disk.

From this commit:
- Added an option to open a wallet by its name. Lookups on default
wallet storage folder by default;
- Wallet data and name are now globally available from mutexes;
- Generated wallets are never opened by default, only
generated. You'll need to open them manually after their creation;

+++ internetz/src/wallet/mod.rs
@@ -14,7 +14,7 @@ use std::fs::File;
 #[derive(Serialize, Deserialize, Clone)]
 pub struct Wallet {
     pub addresses:   Vec<String>,
-    pub balances:    Vec<i64>,
+    pub balances:    Vec<f64>,
     pub last_height: usize,
     pub pubkeys:     Vec<Vec<u8>>,
     privkeys:        Vec<String>, // REALLY REALLY REALLY BAD IDEA. MUST REPLACE.
@@ -53,7 +53,7 @@ impl Wallet {
             // Generate binary addresss
             let binaddr = Wallet::generate_binaddr(&wallet.pubkeys[i]);
             wallet.addresses.push(Wallet::generate_address(&binaddr));
-            wallet.balances.push(0);
+            wallet.balances.push(0.0);
         }
 
         Ok(wallet)
